requirements.txt
requests>=2.28
cryptography>=41.0
python-dotenv>=1.0
flask>=3.0
anthropic>=0.40
